project/requirements.txt
fastapi
fastapi_jwt
mangum==0.10.0
passlib
pydantic
PyJWT
pymysql
requests
uvicorn
pydantic[email]
pandas
gunicorn
snowflake-connector-python[pandas]
python-dotenv
boto3
scikit-image
opencv-python
numpy
pytz
amazon-textract-textractor[pdf,torch]
